requirements/base.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file requirements/base.txt requirements/base.in
#
certifi==2018.4.16        # via requests
chardet==3.0.4            # via requests
click==6.7                # via pip-tools
coreapi==2.3.3            # via drf-yasg, openapi-codec
coreschema==0.0.4         # via coreapi, drf-yasg
django-appconf==1.0.2     # via django-axes
django-axes==4.4.0
django-choices==1.6.1
django-cors-middleware==1.3.1
django-filter==2.0
django-ipware==2.1.0      # via django-axes
django-markup==1.3
django==2.0.13
djangorestframework-camel-case==0.2.0
djangorestframework-filters==0.10.2
djangorestframework-gis==0.13
djangorestframework==3.8.2
drf-nested-routers==0.90.2  # via vng-api-common, zds-schema
drf-writable-nested==0.4.3
drf-yasg==1.8.0
first==2.0.1              # via pip-tools
future==0.16.0            # via drf-yasg
gemma-zds-client==0.9.3   # via vng-api-common, zds-schema
idna==2.6                 # via requests
inflection==0.3.1         # via drf-yasg
iso-639==0.4.5            # via vng-api-common, zds-schema
isodate==0.6.0            # via vng-api-common, zds-schema
itypes==1.1.0             # via coreapi
jinja2==2.10              # via coreschema
markdown==3.0.1
markupsafe==1.0           # via jinja2
openapi-codec==1.3.2      # via drf-yasg
oyaml==0.7                # via vng-api-common, zds-schema
pip-tools==2.0.2
psycopg2-binary==2.7.5
pyjwt==1.6.4              # via gemma-zds-client, vng-api-common, zds-schema
python-dateutil==2.7.3
python-dotenv==0.8.2
pytz==2018.4
pyyaml==3.12              # via gemma-zds-client, oyaml, vng-api-common, zds-schema
raven==6.9.0
requests==2.21.0          # via coreapi, gemma-zds-client, vng-api-common, zds-schema
ruamel.yaml==0.15.37      # via drf-yasg
six==1.11.0               # via django-markup, djangorestframework-gis, drf-yasg, isodate, pip-tools, python-dateutil
unidecode==1.0.22         # via vng-api-common, zds-schema
uritemplate==3.0.0        # via coreapi, drf-yasg
urllib3==1.24.1           # via requests
vng-api-common==0.30.0
zds-schema==0.22.0
